[refactor] Navigation Transition 리팩토링
1. Navigation Transition 애니메이션을 보강
(2. 신고하기 버튼 활성화 버그 수정)
3. NavHost의 composable 내부 composable 들 전부 '스크린이름+Route'로 래핑
