+++ .idea/deploymentTargetDropDown.xml
@@ -2,13 +2,10 @@
 <project version="4">
   <component name="deploymentTargetDropDown">
     <value>
-      <entry key="KeyboardToolbarSamplePreview">
-        <State />
-      </entry>
-      <entry key="NavigationTransitionSamplePreview">
-        <State />
-      </entry>
       <entry key="SampleMainNavigationPreview">
+        <State />
+      </entry>
+      <entry key="SampleNavigationPreview">
         <State />
       </entry>
       <entry key="app">
