+++ app/src/main/java/com/team/flip/navigation/MainNavigation.kt
@@ -1,31 +1,20 @@
 package com.team.flip.navigation
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.team.designsystem.theme.FlipTheme
+import androidx.navigation.compose.rememberNavController
 import com.team.designsystem.theme.FlipTransitionDirection
 import com.team.designsystem.theme.FlipTransitionObject
-import com.team.flip.R
 import com.team.flip.navigation.bottom_nav.FlipBottomNavigationBar
 import com.team.presentation.NavigationItem
 import com.team.presentation.ScreenItem
-import com.team.presentation.editcategories.view.EditMyCategoriesScreen
-import com.team.presentation.editcategories.viewmodel.EditMyCategoriesViewModel
+import com.team.presentation.editcategories.EditCategoriesRoute
 
 /**
  * Flip 의 메인 네비게이션
@@ -34,7 +23,6 @@ import com.team.presentation.editcategories.viewmodel.EditMyCategoriesViewModel
 fun MainNavigation(
     modifier: Modifier = Modifier,
     mainNavController: NavHostController,
-    bottomNavController: NavHostController,
     deleteToken: () -> Unit,
 ) {
 
@@ -43,11 +31,13 @@ fun MainNavigation(
         navController = mainNavController,
         startDestination = NavigationItem.BOTTOM_NAV.name,
         enterTransition = { FlipTransitionObject.enterTransition(FlipTransitionDirection.Left) },
-        exitTransition = { FlipTransitionObject.exitTransition(FlipTransitionDirection.Left) },
         popEnterTransition = { FlipTransitionObject.enterTransition(FlipTransitionDirection.Left) },
-        popExitTransition = {FlipTransitionObject.exitTransition(FlipTransitionDirection.Left) }
+        exitTransition = { FlipTransitionObject.exitTransition(FlipTransitionDirection.Left) },
+        popExitTransition = { FlipTransitionObject.exitTransition(FlipTransitionDirection.Left) }
     ) {
         composable(NavigationItem.BOTTOM_NAV.name) {
+
+            val bottomNavController = rememberNavController()
 
             Scaffold(
                 modifier = Modifier
@@ -77,41 +67,8 @@ fun MainNavigation(
             popExitTransition = { FlipTransitionObject.exitTransition(FlipTransitionDirection.Right) },
         ) {
 
-            val context = LocalContext.current
-            val editMyCategoriesViewModel: EditMyCategoriesViewModel = hiltViewModel()
-            val myCategoriesState by editMyCategoriesViewModel.myCategoriesState.collectAsStateWithLifecycle()
-            val nicknameState by editMyCategoriesViewModel.nicknameState.collectAsStateWithLifecycle()
-            val speechBubbleState by editMyCategoriesViewModel.speechBubbleState.collectAsStateWithLifecycle()
-            val myCategoriesUpdateState by editMyCategoriesViewModel.myCategoriesUpdateState.collectAsStateWithLifecycle()
-
-            val nickname = nicknameState.ifEmpty {
-                context.resources.getString(R.string.placeholder_nickname)
-            }
-
-            LaunchedEffect(myCategoriesUpdateState) {
-                if (myCategoriesUpdateState.success) {
-                    mainNavController.popBackStack()
-                }
-            }
-
-            EditMyCategoriesScreen(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(FlipTheme.colors.white)
-                    .statusBarsPadding()
-                    .navigationBarsPadding()
-                    .padding(bottom = 26.dp),
-                nickname = nickname,
-                myCategoriesState = myCategoriesState,
-                speechBubbleState = speechBubbleState,
-                myCategoriesUpdateState = myCategoriesUpdateState,
-                updateMyCategories = editMyCategoriesViewModel::updateMyCategories,
-                uiEvent = editMyCategoriesViewModel::onUiEvent,
-                onBackPress = {
-                    //TODO 실행 시 HomeScreen 으로 돌아가는데 이 부분에서 딜레이 발생, 해결 필요
-                    // (HomeScreen, HomeViewModel 체크)
-                    mainNavController.popBackStack()
-                }
+            EditCategoriesRoute(
+                popBackStack = { mainNavController.popBackStack() }
             )
         }
     }
